pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powerpoint-mcp"
version = "0.1.0"
description = "MCP server for PowerPoint manipulation using python-pptx"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Juan Ocampo"}
]
keywords = ["mcp", "powerpoint", "pptx", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-pptx>=0.6.21,<0.7",
    "mcp>=1.0.0,<2.0",
    "Pillow>=10.0.0,<11.0",
]

[project.optional-dependencies]
icons = [
    "cairosvg>=2.7.0,<3.0",
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
